requirements.txt
openai-whisper
speechbrain
soundfile
numpy
scipy
